Create Modeling strategy
Our strategy will consist of transforming the work_data frame, typically by adding columns to it at each step. In order to do this, we will create transform functions that take a frame and produce a new frame with altered data.

In particular, we need a function that can produce a transform from an arbitrary regression model. The transform will add model responses as a new column of the frame. Model responses are cross-validated, so they can be used as inputs of other models. The transform is also able to produce output for data that was not used in training.