Create Decorators & data selection
We will start by limiting the data we will use to train models and obtain candidates. In preliminary analysis we found that many potential candidates appear to be below the main sequence in an H-R diagram, but upon further examination, it was clear these are stars with high parallax error. Additionally, model error is distance-dependent.