Create Model variables
Let's define the variables we will use to train regression models. Primarily, we will use "color index" features, which are simply differences between magnitudes observed with different photometric filters. RAVE DR5 also provides metal abundance and metallicity features that seem to help.

Theoretically, the difference between the magnitude of a star and its absolute magnitude is a linear function of the logarithm of the star's distance. But there's an additional extinction component that is roughly dependent on distance. That's why we include both the distance and the logarithm of the distance as variables.