Create Linear modelling
Our first regressor will be linear. Note that we will be modeling the G magnitude of a star, as provided by Gaia DR2. The reader can change the LABEL_COLUMN variable below to model, say, RP or BP magnitudes. There is just one function later in the analysis, called separation_y, that is visually derived and would need to be adjusted accordingly.