Update Blend
It's valid to use the response of a model as a variable of another model, so long as the response is out-of-bag. Our responses are cross-validated, so they should work. If they didn't, that would be evident once we look at the 'test' data we set aside.

We will use linear regression to produce the blend model. A blend does not have to be linear, but in this case that appears to be the best choice.