Create Candidate selection
t's not obvious how candidates should be selected. One possibility is to check how the error distribution compares to a normal distribution. The problem with this approach is that real-world data does not have a normal distribution. Instead, it's typically long-tailed. Outliers are to be expected and they are not indicative that we've found anything "unnatural", nor are non-outliers necessarily "natural".